+++ pom.xml
@@ -2,7 +2,7 @@
          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
-    <groupId>com.csv4pojo</groupId>
+    <groupId>org.csv4pojoparser</groupId>
     <artifactId>csv4pojo</artifactId>
     <version>2.0.0</version>
     <name>CSVUtil</name>
