+++ src/main/java/org/csv4pojoparser/util/CSVWriter.java
@@ -1,4 +1,4 @@
-package com.csv4pojo.util;
+package org.csv4pojoparser.util;
 
 import java.io.Writer;
 import java.util.List;
